Working version of search components
