+++ .vscode/launch.json
@@ -5,7 +5,7 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "type": "edge",
+      "type": "msedge",
       "request": "launch",
       "name": "Launch Edge Canary against localhost",
       "url": "http://localhost:3000",
@@ -13,7 +13,7 @@
       "version": "canary"
     },
     {
-      "type": "edge",
+      "type": "msedge",
       "request": "launch",
       "name": "Launch Edge Dev against localhost",
       "url": "http://localhost:3000",
