+++ package.json
@@ -68,7 +68,9 @@
     "storybook:deploy": "npm run storybook:build && storybook-to-ghpages -e storybook-static",
     "setLicense": "gulp setLicense",
     "test": "jest",
-    "version:tsc": "tsc -v"
+    "version:tsc": "tsc -v",
+    "wtr": "wtr",
+    "wtr:watch": "wtr --watch"
   },
   "storybook-deployer": {
     "gitUsername": "@microsoft/mgt",
@@ -109,6 +111,7 @@
     "@web/dev-server": "^0.1.10",
     "@web/test-runner": "0.15.0",
     "@web/dev-server-esbuild": "0.3.3",
+    "@web/test-runner-playwright": "0.9.0",
     "@webcomponents/webcomponentsjs": "^2.5.0",
     "babel-loader": "^8.2.1",
     "core-js": "^3.7.0",
@@ -172,7 +175,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "npm run prettier:check && npm run lint"
+      "pre-commit": "npm run prettier:check"
     }
   },
   "resolutions": {
