--- packages/mgt-components/src/components/mgt-search-verticals/mgt-search-verticals.scss
@@ -1,1 +1,0 @@
-@import '../../../../../node_modules/office-ui-fabric-core/dist/sass/References';
