+++ packages/mgt-components/src/components/preview/index.ts
@@ -1,2 +1,6 @@
 export * from './mgt-search-box/mgt-search-box';
+//export * from './mgt-search-results-old/mgt-search-results';
 export * from './mgt-search-results/mgt-search-results';
+export * from './mgt-search-filters/mgt-search-filters';
+export * from './mgt-search-verticals/mgt-search-verticals';
+export { IMicrosoftSearchQuery, IMicrosoftSearchService } from '@microsoft/mgt-element';
