+++ packages/mgt-components/src/components/preview/mgt-search-results-old/mgt-search-results.scss
