+++ packages/mgt-components/src/components/preview/mgt-search-verticals/mgt-search-verticals.scss
@@ -1,0 +1,1 @@
+@import '../../../../../../node_modules/office-ui-fabric-core/dist/sass/References';
