+++ packages/mgt-components/src/components/preview/mgt-search-verticals/strings.fr-fr.ts
