+++ packages/mgt-components/src/components/preview/mgt-search-verticals/tests/mgt-search-verticals.test.ts
