+++ packages/mgt-components/src/components/preview/mgt-search-verticals/tests/mocks.ts
