+++ packages/mgt-element/src/ComponentElements.ts
@@ -1,0 +1,3 @@
+export enum ComponentElements {
+  MgtAdaptiveCard = 'mgt-adaptive-card'
+}
